Add `alterF` function.
As per the Data.Map function of the same name, this is a targeted update of a
value in the DMap with a function that returns some `Functor`.

This should be _O (log n)_ as per the `alter` function, but unsure if the
`Functor` that is included could spice that up.

Should I bring across similar documentation to what exists on Data.Map for this
function?

[`alterF` docs](https://hackage.haskell.org/package/containers-0.5.10.2/docs/Data-Map-Strict.html#v:alterF)