Limit problematic use of -trust=* to GHC 7.2.*
SafeHaskell was known to be broken in GHC 7.2 and that's what this hack is for.

However,  using GHC's `-trust=` breaks in various environments and breaks the
`build-depends`-abstraction; I've observed indeterministically breaking builds
because of this.

Using `-trust` on the consumer side is most likely the wrong way anyway, as this
is something that needs to be addressed in `dependent-map`. Future versions of
cabal may actually loudly complain about using `-trust` flags in `ghc-options`.